moved event listeners to control .init() methods
