+++ src/chart/AETable.js
@@ -6,7 +6,6 @@ import { redraw } from './AETable/redraw';
 import { wipe } from './AETable/wipe';
 import { prepareData } from './AETable/prepareData';
 import { init } from './AETable/init';
-import { eventListeners } from './AETable/eventListeners';
 import { toggleRows } from './AETable/toggleRows';
 
 export const AETable =
@@ -14,5 +13,4 @@ export const AETable =
     ,wipe: wipe
     ,prepareData: prepareData
     ,init: init
-    ,eventListeners: eventListeners
     ,toggleRows: toggleRows};
