+++ src/chart/init.js
@@ -72,9 +72,5 @@ export function init(data) {
   //Initialize adverse event eplorer.
     this.layout();
     this.controls.init(this);
-    this.eventListeners.rateFilter(this, this.wrap);
-    this.eventListeners.summaryControl(this, this.wrap, data, settings.variables, settings);
-    this.eventListeners.search(this, this.wrap, data, settings.variables, settings);
-    this.eventListeners.customFilters(this, this.wrap, data, settings.variables, settings);
     this.AETable.redraw(this, this.wrap, data, settings.variables, settings)
 }
